requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
PyMySQL==1.1.1
python-telegram-bot==13.15
pytz==2025.1
requests==2.32.3
six==1.17.0
tornado==6.1
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
